Trying to fix probleb with shadows and instanced mesh

+++ src/objects/InstancedMesh.js
@@ -14,7 +14,7 @@ function InstancedMesh( geometry, material, count ) {
 
 	Mesh.call( this, geometry, material );
 
-	this.instanceMatrix = new BufferAttribute( new Float64Array( count * 16 ), 16 );
+	this.instanceMatrix = new BufferAttribute( new Float32Array( count * 16 ), 16 );
 	this.instanceMatrixModelView = new BufferAttribute( new Float32Array( count * 16 ), 16 );
 	this.instanceMatrixModelView.setUsage(DynamicDrawUsage);
 	this.instanceColor = null;
@@ -32,6 +32,7 @@ InstancedMesh.prototype = Object.assign( Object.create( Mesh.prototype ), {
 	isInstancedMesh: true,
 
 	updateMatrixes: function ( matrix ) {
+		// console.log('call update!!');
 		const mat4 = new Matrix4();
 		for (let index = 0; index < this.count; index++) {
 			mat4.identity();
